requirements.txt
requests>=2.25.0
matplotlib>=3.5.0
numpy>=1.20.0
Pillow>=8.0.0
